requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
cloudinary==1.39.1
Django==4.0.4
django-cloudinary-storage==0.3.0
django-environ==0.11.2
gunicorn==20.1.0
idna==3.3
packaging==23.1
Pillow==9.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2022.1
requests==2.28.1
setuptools==70.3.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.6
whitenoise==6.1.0
